Simple prototype up and running
Started with a simple prototype that creates a visualization using pygame and
ships it to another rpcHugs man which visualizes it. The next step is to
distribute it to lots of displays. Unsure if the current implementation will
run as smoothly as we want it to, but what the heck.